feat: WI-0651 admin approval-executions section file split
