+++ ROADMAP.md
@@ -932,3 +932,4 @@ Phase 8: Extensions (ATS, performance, expenses, analytics)
 - WI-0648 admin people related-workspaces locale normalization (localized ko/en labels for non-devtools related-workspaces panel and aria-label alignment on /admin/people + e2e-wi0648 regression guard)
 - WI-0649 admin approval-executions line-budget decomposition (extract work-conditions/summary/escalation/list/history/logs panels into page-sections with helper/type modules and keep devtools-log gate + dedicated workspace quick-jumps + e2e-wi0649 regression guard)
 - WI-0650 employee attendance prefill effect extraction + line-budget recovery (extract schedule-context correction prefill useEffect from /employee/page.tsx into page-attendance-prefill-effect hook, preserve auto-target selection behavior, and keep employee/page.tsx <=500 lines with e2e-wi0650 guard)
+- WI-0651 admin approval-executions section file split (split page-sections.tsx into work-conditions / summary-escalation / queue section modules, keep each file <=300 lines, and add e2e-wi0651 split+line-budget guard)
